SearchTextFieldをUIコンポーネントに切り出し
